refactor: move idle_logout to utils.py
